Refactor TaskComponent to be less complex and put the interior in another component
